src/requirements.txt
pycryptodome
argon2_cffi
click
